requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
Django==2.2.4
django-background-tasks==1.2.0
django-compat==1.0.15
django-sendsms==0.3.1
django-tables2==2.1.0
django-tags-input==4.6.0
djangorestframework==3.10.2
drawnow==0.72.0
future==0.17.1
gTTS==2.0.3
gTTS-token==1.1.3
html5lib==1.0.1
idna==2.8
iso8601==0.1.12
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
MouseInfo==0.0.4
numpy==1.17.1
opencv-python==4.1.0.25
pandas==0.25.1
Pillow==6.1.0
PyAutoGUI==0.9.47
pygame==1.9.6
PyGetWindow==0.0.7
Pygments==2.4.2
PyMsgBox==1.0.7
pyparsing==2.4.2
PyPDF2==1.26.0
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.22
python-dateutil==2.8.0
PyTweening==1.0.3
pytz==2019.2
PyYAML==5.1.2
reportlab==3.5.23
requests==2.22.0
selenium==3.141.0
serial==0.0.97
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
urllib3==1.25.3
virtualenv==16.7.4
webencodings==0.5.1
xhtml2pdf==0.2.3
